+++ packages/cli/src/lib/CodeGenerator.ts
@@ -1,43 +1,46 @@
-import { SchemaComposer } from "./SchemaComposer";
-import { Project } from "./project";
 import {
   step,
   withSpinner,
   isTypescriptFile,
-  loadTsNode,
-  manifestLanguageToBindLanguage,
-} from "./helpers";
-import { intlMsg } from "./intl";
+  importTypescriptModule,
+  web3apiManifestLanguages,
+  isWeb3ApiManifestLanguage,
+  web3apiManifestLanguageToBindLanguage,
+  pluginManifestLanguages,
+  isPluginManifestLanguage,
+  pluginManifestLanguageToBindLanguage,
+  appManifestLanguages,
+  isAppManifestLanguage,
+  appManifestLanguageToBindLanguage,
+  Project,
+  AnyManifest,
+  SchemaComposer,
+  intlMsg,
+} from "./";
 
 import { TypeInfo } from "@web3api/schema-parse";
 import {
   OutputDirectory,
   writeDirectory,
   bindSchema,
+  BindLanguage,
+  GenerateBindingFn,
 } from "@web3api/schema-bind";
 import path from "path";
 import fs, { readFileSync } from "fs";
 import * as gluegun from "gluegun";
 import { Ora } from "ora";
 import Mustache from "mustache";
-
-export interface CustomScriptConfig {
-  typeInfo: TypeInfo;
-  generate: (templatePath: string, config: unknown) => string;
-}
+import rimraf from "rimraf";
 
 export { OutputDirectory };
 
-export type CustomScriptRunFn = (
-  output: OutputDirectory,
-  config: CustomScriptConfig
-) => void;
-
 export interface CodeGeneratorConfig {
   outputDir: string;
-  project: Project;
+  project: Project<AnyManifest>;
   schemaComposer: SchemaComposer;
   customScript?: string;
+  mustacheView?: Record<string, unknown>;
 }
 
 export class CodeGenerator {
@@ -45,10 +48,10 @@ export class CodeGenerator {
 
   constructor(private _config: CodeGeneratorConfig) {}
 
-  public async generate(): Promise<boolean> {
+  public async generate(config?: Record<string, unknown>): Promise<boolean> {
     try {
       // Compile the API
-      await this._generateCode();
+      await this._generateCode(config);
 
       return true;
     } catch (e) {
@@ -57,18 +60,36 @@ export class CodeGenerator {
     }
   }
 
-  private async _generateCode() {
+  private async _generateCode(config?: Record<string, unknown>) {
     const { schemaComposer, project } = this._config;
 
     const run = async (spinner?: Ora) => {
-      const bindLanguage = manifestLanguageToBindLanguage(
-        await project.getManifestLanguage()
-      );
+      const language = await project.getManifestLanguage();
+      let bindLanguage: BindLanguage | undefined;
 
-      // Make sure that the output dir exists, if not create a new one
-      if (!fs.existsSync(this._config.outputDir)) {
-        fs.mkdirSync(this._config.outputDir);
+      if (isWeb3ApiManifestLanguage(language)) {
+        bindLanguage = web3apiManifestLanguageToBindLanguage(language);
+      } else if (isPluginManifestLanguage(language)) {
+        bindLanguage = pluginManifestLanguageToBindLanguage(language);
+      } else if (isAppManifestLanguage(language)) {
+        bindLanguage = appManifestLanguageToBindLanguage(language);
       }
+
+      if (!bindLanguage) {
+        throw Error(
+          intlMsg.lib_language_unsupportedManifestLanguage({
+            language: language,
+            supported: [
+              ...Object.keys(web3apiManifestLanguages),
+              ...Object.keys(pluginManifestLanguages),
+              ...Object.keys(appManifestLanguages),
+            ].join(", "),
+          })
+        );
+      }
+
+      // Make sure the output dir is reset
+      this._resetDir(this._config.outputDir);
 
       // Get the fully composed schema
       const composed = await schemaComposer.getComposedSchemas();
@@ -88,28 +109,32 @@ export class CodeGenerator {
         const output: OutputDirectory = {
           entries: [],
         };
-
-        if (isTypescriptFile(this._config.customScript)) {
-          loadTsNode();
-        }
+        const customScript = this._config.customScript;
 
         // Check the generation file if it has the proper run() method
-        // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports, @typescript-eslint/naming-convention
-        const generator = await require(this._config.customScript);
+        // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports
+        const generator = isTypescriptFile(customScript)
+          ? await importTypescriptModule(customScript)
+          : // eslint-disable-next-line @typescript-eslint/no-require-imports
+            await require(customScript);
+
         if (!generator) {
           throw Error(intlMsg.lib_codeGenerator_wrongGenFile());
         }
 
-        const { run } = generator as { run: CustomScriptRunFn };
-        if (!run) {
-          throw Error(intlMsg.lib_codeGenerator_noRunMethod());
+        const { generateBinding } = generator as {
+          generateBinding: GenerateBindingFn;
+        };
+        if (!generateBinding) {
+          throw Error(intlMsg.lib_codeGenerator_nogenerateBindingMethod());
         }
 
-        await run(output, {
+        await generateBinding(
+          output,
           typeInfo,
-          generate: (templatePath: string, config: unknown) =>
-            this._generateTemplate(templatePath, config, spinner),
-        });
+          this._schema || "",
+          config || {}
+        );
 
         writeDirectory(this._config.outputDir, output, (templatePath: string) =>
           this._generateTemplate(templatePath, typeInfo, spinner)
@@ -119,6 +144,7 @@ export class CodeGenerator {
           combined: {
             typeInfo: composed.combined?.typeInfo as TypeInfo,
             schema: composed.combined?.schema as string,
+            config,
             outputDirAbs: "",
           },
           bindLanguage,
@@ -143,6 +169,14 @@ export class CodeGenerator {
         }
       );
     }
+  }
+
+  private _resetDir(dir: string) {
+    if (fs.existsSync(dir)) {
+      rimraf.sync(dir);
+    }
+
+    fs.mkdirSync(dir, { recursive: true });
   }
 
   private _generateTemplate(
@@ -173,6 +207,7 @@ export class CodeGenerator {
     let content = Mustache.render(template.toString(), {
       ...types,
       schema: this._schema,
+      ...this._config.mustacheView,
     });
 
     content = `// ${intlMsg.lib_codeGenerator_templateNoModify()}