+++ packages/cli/src/lib/helpers/client.ts
@@ -1,0 +1,52 @@
+import { PluginRegistration } from "@web3api/core-js";
+import { ensPlugin } from "@web3api/ens-plugin-js";
+import { ethereumPlugin } from "@web3api/ethereum-plugin-js";
+import { ipfsPlugin } from "@web3api/ipfs-plugin-js";
+import { Web3ApiClient, defaultIpfsProviders } from "@web3api/client-js";
+
+interface SimpleClientConfig {
+  ensAddress?: string;
+  ethProvider?: string;
+  ipfsProvider?: string;
+}
+
+export function getSimpleClient(config: SimpleClientConfig): Web3ApiClient {
+  const { ensAddress, ethProvider, ipfsProvider } = config;
+  const plugins: PluginRegistration[] = [];
+  if (ensAddress) {
+    plugins.push({
+      uri: "w3://ens/ens.web3api.eth",
+      plugin: ensPlugin({
+        addresses: {
+          testnet: ensAddress,
+        },
+      }),
+    });
+  }
+  if (ethProvider) {
+    plugins.push({
+      uri: "w3://ens/ethereum.web3api.eth",
+      plugin: ethereumPlugin({
+        networks: {
+          testnet: {
+            provider: ethProvider,
+          },
+          mainnet: {
+            provider:
+              "https://mainnet.infura.io/v3/b00b2c2cc09c487685e9fb061256d6a6",
+          },
+        },
+      }),
+    });
+  }
+  if (ipfsProvider) {
+    plugins.push({
+      uri: "w3://ens/ipfs.web3api.eth",
+      plugin: ipfsPlugin({
+        provider: ipfsProvider,
+        fallbackProviders: defaultIpfsProviders,
+      }),
+    });
+  }
+  return new Web3ApiClient({ plugins });
+}